refactor(connector): create graph
