+++ packages/connector/src/index.ts
@@ -1,5 +1,5 @@
 export * from './a-star.js';
-export * from './create-graph.js';
+export * from './create-graph/create-graph.js';
 export * from './rectangle.js';
 export * from './route.js';
 export * from './simplify-path.js';
